Address review feedback on Redis Sentinel support
- Default tight TCP keepalive on the Sentinel data-node pool so a master
  that dies silently (network partition / frozen host) is noticed in
  seconds instead of waiting on the OS-default keepalive. docket disables
  the read timeout for its long blocking reads, and redis-py <8 (which
  docket still supports via redis>=5) defaults to no keepalive at all, so
  a blocked XREADGROUP would otherwise hang indefinitely past failover.
  Reproduced end to end against a real master+replica+sentinel topology:
  the blocked read now raises ConnectionError in ~45s rather than never.
- Move all Sentinel code into a companion _redis_sentinel.py module;
  _redis.py keeps only is_sentinel and the dispatch (1116 -> 922 lines),
  restoring the loq baseline.
- Parse each sentinel member with urlparse(f"//{member}") instead of the
  hand-rolled bracket/port handling (hostnames, ports, default ports,
  bracketed IPv6, and port validation all come for free).
- Detect the sentinel scheme with a string-prefix check and only carve the
  multi-host netloc by hand for sentinel URLs; standalone, cluster, and
  memory URLs go straight through urlparse. Phrase the carving helper's
  docstring timelessly.
- Drop the ?db=-beats-path and query-credential precedence emulation: the
  database comes from the path and master credentials from the userinfo;
  the parse_url funnel for typed pool options stays.
